+++ packages/components/src/Status/Status.test.tsx
@@ -24,49 +24,27 @@
 
  */
 
-export * from './Accordion'
-export * from './ActionList'
-export * from './Avatar'
-export * from './Badge'
-export * from './Banner'
-export * from './Button'
-export * from './Calendar'
-export * from './Card'
-export * from './Chip'
-export * from './DateTimeFormat'
-export * from './Divider'
-export * from './Form'
-export * from './Icon'
-export * from './Layout'
-export * from './Link'
-export * from './List'
-export * from './Menu'
-export * from './Modal'
-export * from './PageSize'
-export * from './Pagination'
-export * from './Popover'
-export * from './Spinner'
-export * from './Status'
-export * from './Sidebar'
-export * from './Table'
-export * from './Tabs'
-export * from './Tooltip'
-export * from './Text'
-export * from './Tree'
-export * from './VisuallyHidden'
+import 'jest-styled-components'
+import React from 'react'
+import { assertSnapshot } from '@looker/components-test-utils'
+import { Status } from './Status'
 
-export * from './utils'
+test('critical Status', () => {
+  assertSnapshot(<Status intent="critical" />)
+})
 
-export { ComponentsProvider } from '@looker/components-providers'
+test('inform Status', () => {
+  assertSnapshot(<Status intent="inform" />)
+})
 
-/** Provided here for backwards compatibility.
- * @TODO - Remove before 1.0
- **/
+test('neutral status', () => {
+  assertSnapshot(<Status intent="neutral" />)
+})
 
-export {
-  GlobalStyle,
-  palette,
-  SpacingSizes,
-  theme,
-  Theme,
-} from '@looker/design-tokens'
+test('positive Status', () => {
+  assertSnapshot(<Status intent="positive" />)
+})
+
+test('warning Status', () => {
+  assertSnapshot(<Status intent="warning" />)
+})